@kooksee 2023-06-08
@PokIsemaine Removing cache is a good idea, but when users customize matchingFunc and the functions in matchingFunc are slow, it still causes the `enforce` to be slow. Of course, this is user's problem